add mechanism to tell gui that item was replaced
before the problem of notifying the visualizer of an item update (with lower version)
was handled by overriding the version of the new item with the version of the old item
which required to write to the shared data in the elderpt thread which was breaking
the type system and wasn't always working reulting in non updated visualizers after
restarting elderpt.

the explicit mechanism removes this problem.